변수 타입 다시 적음
일단 보이는대로 정리했는데... 어딘가 덜 고쳤을 수도 있음

+++ g_board/src/main/java/com/mycompany/myapp/service/BoardService.java
@@ -21,5 +21,5 @@ public interface BoardService {
 	
 	public List<BoardVO> listWithPaging(Criteria cri);
 	
-	public long count();
+	public int count();
 }
